backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Async SQLite (local dev) + PostgreSQL (production)
sqlalchemy[asyncio]>=2.0.40
aiosqlite>=0.19.0
# asyncpg==0.30.0  # Only needed for PostgreSQL production deployments
alembic>=1.14.0

# Auth
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Config / Env
pydantic-settings>=2.7.0
pydantic[email]>=2.10.6
email-validator==2.2.0

# File uploads (CSV)
python-multipart==0.0.20

# Async file I/O
aiofiles==24.1.0

# Utilities
python-dateutil==2.9.0
python-dotenv>=1.0.0
jinja2>=3.1.0
# joblib>=1.3.0
# scikit-learn>=1.3.0
# numpy>=1.24.0

# RAG Chatbot (optional — app starts fine without these)
# Uncomment if you need the /chat endpoint:
# groq>=0.4.0
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
